langgraph-ollama-agent/pyproject.toml
[project]
name = "langgraph-ollama-agent"
version = "0.1.0"
description = "Multi-agent system with LangGraph and Ollama"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "openai-agents==0.0.4",
    "langgraph==0.2.16",
    "langchain-core>=0.2.27",
    "openai>=1.0.0",
    "duckduckgo-search==6.1.4",
    "python-dotenv==1.0.1",
    "pydantic>=2.10",
    "loguru==0.7.2",
    "colorama==0.4.6",
    "httpx==0.27.0",
    "aiofiles==23.2.1",
]

[project.optional-dependencies]
dev = [
    "pytest==8.0.0",
    "pytest-asyncio==0.23.0",
]

jupyter = [
    "jupyter==1.0.0",
    "ipython==8.18.1",
]

dev-tools = [
    "black==23.12.0",
    "flake8==6.1.0",
    "mypy==1.7.1",
    "isort==5.13.2",
]
